refactor rsc-plugin
